requirements.txt
mysql-connector-python
bcrypt
reportlab